# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Single sign-on (сервис единого входа)",
        "title": "SSO",
        "contact": {
            "name": "Alnovi",
            "url": "https://github.com/alnovi"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/alnovi/sso/refs/heads/master/LICENSE.md"
        },
        "version": "1.0.0"
    },
    "host": "http://localhost:8081",
    "paths": {
...
        "/api/users": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Список пользователей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Список пользователей",
                "operationId": "UsersList",
                "responses": {
                    "200": {
                        "description": "Информация о пользователях",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.User"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Добавление нового пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Добавление пользователя",
                "operationId": "UsersCreate",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/response.User"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Информация о пользователе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Пользователь",
                "operationId": "UsersGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/response.User"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Изменение пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Изменение пользователя",
                "operationId": "UsersUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/response.User"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Удаление пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Удаление пользователя",
                "operationId": "UsersDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/response.User"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/users/{id}/clients": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Список приложений и права доступа для пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Список клиентов пользователя",
                "operationId": "UsersClients",
                "responses": {
                    "200": {
                        "description": "Информация о клиентах и правах доступа",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.ClientRole"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users/{id}/restore": {
            "post": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Восстановление удаленного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Восстановление пользователя",
                "operationId": "UsersRestore",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/response.User"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/users/{uid}/clients/{cid}": {
            "post": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Изменение роли доступа пользователя в приложении",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Users"
                ],
                "summary": "Изменение роли пользователя",
                "operationId": "UsersUpdateRole",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор клиента",
                        "name": "сid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Роль пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateUserRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateUser": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 24,
                    "minLength": 5
                }
            }
        },
...
        "request.UpdateUser": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 24,
                    "minLength": 5
                }
            }
        },
        "request.UpdateUserRole": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "guest",
                        "user",
                        "manager",
                        "admin"
                    ]
                }
            }
        },
...
        "response.ClientRole": {
            "type": "object",
            "properties": {
                "callback": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_system": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT-Access": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}